requirements.txt
agentic-doc
pandas==2.2.0
python-dotenv==1.0.1
google-generativeai>=0.3.0
